# openapi_schema.yaml
openapi: 3.1.0
info:
  title: China Securities Data Proxy
  description: >-
    REST API-обёртка над AKShare для доступа к финансовой отчётности,
    дивидендам, дневным котировкам и валютным курсам China A-share (SSE/SZSE)
    и Hong Kong (HKEX) тикеров.
  version: 1.0.0
servers:
  - url: https://china-securities-proxy.onrender.com
paths:
...
  /income:
    get:
      operationId: getIncomeStatement
      summary: Отчёт о прибылях и убытках A-share тикера
      parameters:
        - name: ts_code
          in: query
          required: true
          description: Тикер (суффиксы .SH/.SZ/.HK будут очищены автоматически)
          schema:
            type: string
        - name: periods
          in: query
          required: false
          description: >-
            Сколько последних отчётных периодов вернуть (по умолчанию 8).
            Полная история эмитента не помещается в лимит ответа Action,
            поэтому выдача обрезается с конца — самые свежие периоды всегда
            попадают в ответ. Увеличивай только если действительно нужны
            более старые периоды.
          schema:
            type: integer
            default: 8
      responses:
        "200":
          description: Успешный ответ с данными или пустым результатом
          content:
            application/json:
              schema:
                type: object
                properties:
                  periods_returned:
                    type: integer
                    description: Сколько отчётных периодов фактически вернулось
                  periods_available:
                    type: integer
                    description: >-
                      Сколько всего периодов есть у источника. Если больше
                      periods_returned, выдача урезана — скажи об этом
                      пользователю, а не выдавай её за всю историю.
                  data:
                    type: array
                    items:
                      type: object
                  note:
                    type: string
        "502":
          description: Ошибка при обращении к источнику данных (AKShare)
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
  /balancesheet:
    get:
      operationId: getBalanceSheet
      summary: Баланс A-share тикера
      parameters:
        - name: ts_code
          in: query
          required: true
          description: Тикер (суффиксы .SH/.SZ/.HK будут очищены автоматически)
          schema:
            type: string
        - name: periods
          in: query
          required: false
          description: >-
            Сколько последних отчётных периодов вернуть (по умолчанию 8).
            Полная история эмитента не помещается в лимит ответа Action,
            поэтому выдача обрезается с конца — самые свежие периоды всегда
            попадают в ответ. Увеличивай только если действительно нужны
            более старые периоды.
          schema:
            type: integer
            default: 8
      responses:
        "200":
          description: Успешный ответ с данными или пустым результатом
          content:
            application/json:
              schema:
                type: object
                properties:
                  periods_returned:
                    type: integer
                    description: Сколько отчётных периодов фактически вернулось
                  periods_available:
                    type: integer
                    description: >-
                      Сколько всего периодов есть у источника. Если больше
                      periods_returned, выдача урезана — скажи об этом
                      пользователю, а не выдавай её за всю историю.
                  data:
                    type: array
                    items:
                      type: object
                  note:
                    type: string
        "502":
          description: Ошибка при обращении к источнику данных (AKShare)
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
  /cashflow:
    get:
      operationId: getCashFlow
      summary: Отчёт о движении денежных средств A-share тикера
      parameters:
        - name: ts_code
          in: query
          required: true
          description: Тикер (суффиксы .SH/.SZ/.HK будут очищены автоматически)
          schema:
            type: string
        - name: periods
          in: query
          required: false
          description: >-
            Сколько последних отчётных периодов вернуть (по умолчанию 8).
            Полная история эмитента не помещается в лимит ответа Action,
            поэтому выдача обрезается с конца — самые свежие периоды всегда
            попадают в ответ. Увеличивай только если действительно нужны
            более старые периоды.
          schema:
            type: integer
            default: 8
      responses:
        "200":
          description: Успешный ответ с данными или пустым результатом
          content:
            application/json:
              schema:
                type: object
                properties:
                  periods_returned:
                    type: integer
                    description: Сколько отчётных периодов фактически вернулось
                  periods_available:
                    type: integer
                    description: >-
                      Сколько всего периодов есть у источника. Если больше
                      periods_returned, выдача урезана — скажи об этом
                      пользователю, а не выдавай её за всю историю.
                  data:
                    type: array
                    items:
                      type: object
                  note:
                    type: string
        "502":
          description: Ошибка при обращении к источнику данных (AKShare)
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
  /dividend:
    get:
      operationId: getDividendHistory
      summary: История дивидендов A-share тикера
      parameters:
        - name: ts_code
          in: query
          required: true
          description: Тикер (суффиксы .SH/.SZ/.HK будут очищены автоматически)
          schema:
            type: string
      responses:
        "200":
          description: Успешный ответ с данными или пустым результатом
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  note:
                    type: string
        "502":
          description: Ошибка при обращении к источнику данных (AKShare)
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
  /daily:
    get:
      operationId: getDailyQuotesAShare
      summary: Дневные котировки A-share тикера
      parameters:
        - name: ts_code
          in: query
          required: true
          description: Тикер (суффиксы .SH/.SZ/.HK будут очищены автоматически)
          schema:
            type: string
        - name: start_date
          in: query
          required: true
          description: Дата начала периода в формате YYYY-MM-DD
          schema:
            type: string
        - name: end_date
          in: query
          required: true
          description: Дата конца периода в формате YYYY-MM-DD
          schema:
            type: string
      responses:
        "200":
          description: Успешный ответ с данными или пустым результатом
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  note:
                    type: string
        "502":
          description: Ошибка при обращении к источнику данных (AKShare)
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
  /hk_daily:
    get:
      operationId: getDailyQuotesHK
      summary: >-
        Дневные котировки HK тикера (формат тикера без суффикса .HK, напр.
        "00267"). Для периодов от ~2 лет и больше используйте interval=weekly
        или monthly/yearly — daily-ответ на несколько лет может быть очень
        большим и медленным.
      parameters:
        - name: ts_code
          in: query
          required: true
          description: Тикер (суффиксы .SH/.SZ/.HK будут очищены автоматически)
          schema:
            type: string
        - name: start_date
          in: query
          required: true
          description: Дата начала периода в формате YYYY-MM-DD
          schema:
            type: string
        - name: end_date
          in: query
          required: true
          description: Дата конца периода в формате YYYY-MM-DD
          schema:
            type: string
        - name: interval
          in: query
          required: false
          description: >-
            Агрегация OHLCV: "daily" (по умолчанию, вся история по дням —
            для многолетних периодов ответ большой), "weekly", "monthly"
            или "yearly" (одна строка на период: open первого дня, high/low
            экстремумы, close последнего дня, volume/amount — сумма).
          schema:
            type: string
            default: daily
      responses:
        "200":
          description: Успешный ответ с данными или пустым результатом
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  note:
                    type: string
        "502":
          description: Ошибка при обращении к источнику данных (AKShare)
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
  /hk_financial:
    get:
      operationId: getFinancialsHK
      summary: Финансовые показатели HK тикера
      parameters:
        - name: ts_code
          in: query
          required: true
          description: Тикер (суффиксы .SH/.SZ/.HK будут очищены автоматически)
          schema:
            type: string
        - name: periods
          in: query
          required: false
          description: >-
            Сколько последних отчётных периодов вернуть (по умолчанию 8).
            Полная история эмитента не помещается в лимит ответа Action,
            поэтому выдача обрезается с конца — самые свежие периоды всегда
            попадают в ответ. Увеличивай только если действительно нужны
            более старые периоды.
          schema:
            type: integer
            default: 8
      responses:
        "200":
          description: Успешный ответ с данными или пустым результатом
          content:
            application/json:
              schema:
                type: object
                properties:
                  periods_returned:
                    type: integer
                    description: Сколько отчётных периодов фактически вернулось
                  periods_available:
                    type: integer
                    description: >-
                      Сколько всего периодов есть у источника. Если больше
                      periods_returned, выдача урезана — скажи об этом
                      пользователю, а не выдавай её за всю историю.
                  data:
                    type: array
                    items:
                      type: object
                  note:
                    type: string
        "502":
          description: Ошибка при обращении к источнику данных (AKShare)
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
  /fx:
    get:
      operationId: getFxRateHistory
      summary: История официальной котировки Банка Китая для валюты к CNY
      parameters:
        - name: currency
          in: query
          required: false
          description: >-
            Китайское название валюты, напр. "港币" (HKD), "美元" (USD),
            "欧元" (EUR). По умолчанию "港币".
          schema:
            type: string
            default: 港币
        - name: start_date
          in: query
          required: true
          description: Дата начала периода в формате YYYY-MM-DD
          schema:
            type: string
        - name: end_date
          in: query
          required: true
          description: Дата конца периода в формате YYYY-MM-DD
          schema:
            type: string
      responses:
        "200":
          description: Успешный ответ с данными или пустым результатом
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  note:
                    type: string
        "502":
          description: Ошибка при обращении к источнику данных (AKShare)
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
  /fx_cny_hkd_on_date:
    get:
      operationId: getFxRateCnyHkdOnDate
      summary: >-
        Официальная котировка Банка Китая (BOC) CNY/HKD на конкретную дату
        (прямая котировка, не кросс-расчёт)
      parameters:
        - name: date
          in: query
          required: true
          description: Дата в формате YYYY-MM-DD
          schema:
            type: string
      responses:
        "200":
          description: Котировка на дату вместе с указанием источника и методологии
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  note:
                    type: string
                  source:
                    type: string
                  methodology:
                    type: string
        "502":
          description: Ошибка при обращении к источнику данных (AKShare)
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
  /stock_basic:
    get:
      operationId: getStockBasicInfo
      summary: Базовая информация о тикере (A-share)
      parameters:
        - name: ts_code
          in: query
          required: true
          description: Тикер (суффиксы .SH/.SZ/.HK будут очищены автоматически)
          schema:
            type: string
      responses:
        "200":
          description: Успешный ответ с данными или пустым результатом
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  note:
                    type: string
        "502":
          description: Ошибка при обращении к источнику данных (AKShare)
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
...